requirements.txt
# Football Big Data Analytics - Python Requirements

# Core Data Processing
pyspark==3.4.1
pandas==2.1.3
numpy==1.24.3

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Database Drivers
psycopg2-binary==2.9.9
pymongo==4.6.0
elasticsearch==8.11.0

# API Clients
requests==2.31.0
requests-oauthlib==1.3.1

# Streaming & Real-time
apache-airflow==2.7.3
flink==1.17.1

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
xgboost==2.0.2
lightgbm==4.1.0
mlflow==2.9.2

# Data Validation & Quality
great-expectations==0.18.5
pandera==0.17.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# NLP & Sentiment Analysis
nltk==3.8.1
textblob==0.17.1
transformers==4.36.0

# API Development
fastapi==0.105.0
uvicorn==0.25.0
pydantic==2.5.2
graphene==3.3

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
schedule==1.2.0
tqdm==4.66.1

# Jupyter & Notebooks
jupyter==1.0.0
jupyterlab==4.0.9
ipywidgets==8.1.1

# Cloud & Storage
boto3==1.34.10
azure-storage-blob==12.19.0
google-cloud-storage==2.13.0

# Data Formats
pyarrow==14.0.1
avro-python3==1.10.2
